Fixed duplicate migration with schema version 36
