+++ golem/database/database.py
@@ -61,7 +61,7 @@ class GolemSqliteDatabase(peewee.SqliteDatabase):
 
 
 class Database:
-    SCHEMA_VERSION = 39
+    SCHEMA_VERSION = 40
 
     def __init__(self,  # noqa pylint: disable=too-many-arguments
                  db: peewee.Database,
