--- golem/database/schemas/036_task_payment_charged_from_deposit.py
@@ -1,16 +1,0 @@
-# pylint: disable=no-member
-# pylint: disable=unused-argument
-from golem import model
-
-SCHEMA_VERSION = 36
-
-
-def migrate(migrator, database, fake=False, **kwargs):
-    migrator.add_fields(
-        'taskpayment',
-        charged_from_deposit=model.BooleanField(null=True),
-    )
-
-
-def rollback(migrator, database, fake=False, **kwargs):
-    migrator.remove_fields('taskpayment', 'charged_from_deposit')
