+++ golem/database/schemas/037_task_payment_charged_from_deposit.py
@@ -1,17 +1,16 @@
 # pylint: disable=no-member
 # pylint: disable=unused-argument
-# pylint: disable=unused-variable
-
-import peewee as pw
+from golem import model
 
 SCHEMA_VERSION = 37
 
 
 def migrate(migrator, database, fake=False, **kwargs):
     migrator.add_fields(
-        'requestedtask',
-        env_id=pw.CharField(max_length=255, null=True))
+        'taskpayment',
+        charged_from_deposit=model.BooleanField(null=True),
+    )
 
 
 def rollback(migrator, database, fake=False, **kwargs):
-    migrator.remove_fields('requestedtask', 'env_id')
+    migrator.remove_fields('taskpayment', 'charged_from_deposit')
